Move `constitution.jax.isochoric_volumetric_split()`
from `constitution.jax.models.hyperelastic.isochoric_volumetric_split()`; also for tensortrax
