+++ src/felupe/constitution/tensortrax/__init__.py
@@ -1,4 +1,5 @@
 from . import models
+from ._helpers import isochoric_volumetric_split
 from ._hyperelastic import Hyperelastic
 from ._material import Material
 from ._total_lagrange import total_lagrange
@@ -6,6 +7,7 @@ from ._updated_lagrange import updated_lagrange
 
 __all__ = [
     "Hyperelastic",
+    "isochoric_volumetric_split",
     "models",
     "total_lagrange",
     "updated_lagrange",
